+++ simpegPF/__init__.py
@@ -3,4 +3,4 @@ import BaseMag
 import Magnetics
 import BaseGrav
 import Gravity
-import MagneticsIO
+import MagneticsDriver
